Refactor error handling and API services
- Updated `shouldReportError` function to only suppress reporting for 400 errors.
- Modified `createMutationErrorHandler` to prevent duplicate Sentry reporting.
- Removed 401 error filtering from Sentry's `beforeSend` hook.
- Introduced a new `api-client.ts` file to centralize API configuration and request handling.
- Split services into separate files: `auth.service.ts`, `budgets.service.ts`, `categories.service.ts`, `dashboard.service.ts`, `recurring-transactions.service.ts`, and `transactions.service.ts`.
- Enhanced error handling in each service to log errors silently and prevent duplicate notifications.
